JCBC-1023 Ignore FTS tests if search service not found
Motivation
----------
Since FTS is an individual service, we may not be
enabling it always while running integration tests.

Changes
-------
Add ignoreIfSearchServiceNotFound to test context, which tries a search
query and does not continue with the tests if a
ServiceNotAvailableException is returned.

Results
-------
Fts tests will be ignored if search service is not found

Reviewed-on: http://review.couchbase.org/70046
Tested-by: Subhashni Balakrishnan <b.subhashni@gmail.com>

+++ src/integration/java/com/couchbase/client/java/SearchQueryTest.java
@@ -76,17 +76,20 @@ public class SearchQueryTest {
     public static void init() throws InterruptedException {
         ctx = CouchbaseTestContext.builder()
                 .withEnv(DefaultCouchbaseEnvironment.builder()
-                    .mutationTokensEnabled(true))
+                        .mutationTokensEnabled(true))
                 .bucketName("beer-sample")
                 .flushOnInit(false)
                 .adhoc(false)
                 .build()
-                .ignoreIfMissing(CouchbaseFeature.FTS_BETA);
+                .ignoreIfMissing(CouchbaseFeature.FTS_BETA)
+                .ignoreIfSearchServiceNotFound();
     }
 
     @AfterClass
     public static void cleanup() {
-        ctx.destroyBucketAndDisconnect();
+        if (ctx != null) {
+            ctx.destroyBucketAndDisconnect();
+        }
     }
 
     @Test
